refactor: one e2e test file per screen
