+++ test/e2e/fixtures/file.txt
@@ -1,0 +1,1 @@
+hello i am a text file
